move KEEPALIVE_JOB_ID_DELIMITER, is_mining_authorize and tlv_compatible_username to utils and make changes accordingly
