+++ miner-apps/translator/src/lib/utils.rs
@@ -18,14 +18,16 @@ use stratum_apps::{
         },
         sv1_api::{
             client_to_server::{self, Submit},
+            json_rpc,
             server_to_client::Notify,
             utils::HexU32Be,
+            Message,
         },
     },
     utils::types::{ChannelId, DownstreamId},
 };
 
-use tracing::debug;
+use tracing::{debug, warn};
 
 use crate::error::TproxyErrorKind;
 
@@ -246,4 +248,42 @@ pub struct SubmitShareWithChannelId {
     pub version_rolling_mask: Option<HexU32Be>,
     /// The version field from the job, used for validation
     pub job_version: Option<u32>,
+}
+
+/// Delimiter used to separate original job ID from keepalive mutation counter.
+/// Format: `{original_job_id}#{counter}`
+pub(crate) const KEEPALIVE_JOB_ID_DELIMITER: char = '#';
+
+/// Check if Sv1 message is mining.authorize
+pub(crate) fn is_mining_authorize(msg: &Message) -> bool {
+    if let json_rpc::Message::StandardRequest(r) = &msg {
+        r.method == "mining.authorize"
+    } else {
+        false
+    }
+}
+
+/// Truncates a string to [`MAX_USER_IDENTITY_BYTES`], respecting UTF-8 character boundaries.
+///
+/// If the input string exceeds the limit, it is truncated at the last valid UTF-8 character
+/// boundary before or at [`MAX_USER_IDENTITY_BYTES`] and a warning is logged.
+pub(crate) fn tlv_compatible_username(s: &str) -> &str {
+    const MAX_USER_IDENTITY_BYTES: usize = 32;
+    let len = s.len();
+
+    if len <= MAX_USER_IDENTITY_BYTES {
+        return s;
+    }
+    // Find the last valid UTF-8 char boundary at or before MAX_USER_IDENTITY_BYTES
+    let mut end = MAX_USER_IDENTITY_BYTES;
+    while end > 0 && !s.is_char_boundary(end) {
+        end -= 1;
+    }
+    let truncated = &s[..end];
+    warn!(
+        "Username '{}' exceeds {} bytes ({} bytes), truncating to '{}'. \
+         Consider using a shorter username for full visibility on the pool dashboard.",
+        s, MAX_USER_IDENTITY_BYTES, len, truncated
+    );
+    truncated
 }
